backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
flask-openapi3==3.1.3
Flask-SQLAlchemy==3.1.1
openapi3==1.8.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pytest==8.3.3
pytest-flask==1.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
